Fix Chrome warning's design
to look like the real thing

I did this by importing the source of an SSL warning as seen on
chrome://interstitials (and putting back the old customizations,
including interaction tracking and the condition code).

It is now pretty much identical to the real warning (as far as I can tell).