Serve warnings and target page from the same URL
Instead of having a separate /proceed page to which users get
redirected, now, when the user chooses to proceed, the app state is
mutated, the page is reloaded, and the same page now shows the target
website.

The main benefit is that this is more realistic: there's no more awkward
(and noticeable) URL change. Browser history (i.e., the back button)
will also work as expected. Finally, if the user ends up returning to
the page, their decision will have been saved, as with a real warning.

However, the new change requires that the visitor has JavaScript enabled.