Fix check for existing session values
The code checked if values were present in the session with a truthiness
check (! value). However, in JavaScript 0 is falsey, and enum values
in TypeScript are integers, so anytime the value was the first entry in
an enum, the check thought it wasn't set, and re-set it.

This resulted in a number of duplicate events being generated, confusing
potential data analysis.